Fixes for scanning eclipse binary files
For numerical data, a record size of 1000 was assumed, which proves incorrect for many files. fortio_data_fskip() now reads the record size from the binary file and skips ahead based on that. The function is optimized to make no more that 2 fseek calls in ordinary cases (N records of equal size + 1 or 0 records of a smaller size), but falls back to skipping each record individually should the first method fail.